% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-to-atu.R
\name{edw_date_to_atu}
\alias{edw_date_to_atu}
\title{Date to Accumulated Thermal Units (ATUs)}
\usage{
edw_date_to_atu(date, atu, dts, date_time = "DateTime",
  value = "Value")
}
\arguments{
\item{date}{A Date vector.}

\item{atu}{A count of the ATUs}

\item{dts}{A dts object.}

\item{date_time}{A string of the Date column name in dts.}

\item{value}{A string of the temperature column name in dts.}
}
\value{
A Date vector.
}
\description{
Date to Accumulated Thermal Units (ATUs)
}
