% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatch.R
\name{edw_hatch}
\alias{edw_hatch}
\title{Date to Hatch based on Accumulated Thermal Units (ATUs)}
\usage{
edw_hatch(x, dtt = "DateTime", colname = "Value", atus = list(Hatch =
  320L, Emergence = 480L))
}
\arguments{
\item{x}{A dts object.}

\item{dtt}{A string of the name of the column with the Date or POSIXct values.}

\item{colname}{A string of the name of the column with the temperature data.}

\item{atus}{A named list of the daily accumulated thermal units to life stage.}
}
\value{
The modified dts object.
}
\description{
Date to Hatch based on Accumulated Thermal Units (ATUs)
}
