% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn-timing.R
\name{edw_spawn_timing}
\alias{edw_spawn_timing}
\title{Spawn Timing}
\usage{
edw_spawn_timing(date, sd, prop = 0.95)
}
\arguments{
\item{date}{A Date vector of the mean spawn date.}

\item{sd}{A number of the standard deviation in the spawning distribution.}

\item{prop}{A number of the proportion of spawning to consider.}
}
\value{
A dts object where DateTime is the Date, Value is the yearly proportion
and MeanSpawnDate is the mean spawn date.
}
\description{
Assumes that spawning is normally distributed.
There must be no overlap between the spawning associated with each mean spawn date.
}
\examples{
edw_spawn_timing(as.Date(c("2000-03-30", "2001-04-06")), sd = 1)
}
