% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawning.R
\name{edw_spawning}
\alias{edw_spawning}
\title{Spawning}
\usage{
edw_spawning(x, dtt = "DateTime", colname = "Spawning",
  peak = as.Date("1972-06-01"), sd = 7L)
}
\arguments{
\item{x}{A dts data frame.}

\item{dtt}{A string of the name of the column with the DateTime values.}

\item{colname}{A string of the name of the column to save the spawning densities to.}

\item{peak}{A Date vector of the mean spawn dates.}

\item{sd}{A positive integer scalar or vector of the standard deviation in days in the distribution of spawning.}
}
\value{
The dts data frame with a column of the spawning densities.
}
\description{
The spawning densities are calculated using dnorm.
Its important to realize that the densities will need rescaling.
}
