% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{edw_survival_elevation}
\alias{edw_survival_elevation}
\title{Survival by Elevation}
\usage{
edw_survival_elevation(nest, eggs, egg_int = 0L)
}
\arguments{
\item{nest}{A number of the nest elevation.}

\item{eggs}{A double vector of the water elevations during the egg stage.}

\item{egg_int}{A count of the number intervals eggs can survive dewatering.}
}
\value{
A number of the proportion surviving.
}
\description{
Survival by Elevation
}
