% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{edw_survival_elevation2}
\alias{edw_survival_elevation2}
\title{Survival by Elevation and Intervals}
\usage{
edw_survival_elevation2(nest, eggs, alevins, egg_int = 0L, alevin_int = 0L)
}
\arguments{
\item{nest}{A number of the nest elevation.}

\item{eggs}{A double vector of the water elevations during the egg stage.}

\item{alevins}{A double vector of the water elevations during the alevin stage.}

\item{egg_int}{A count of the number intervals eggs can survive dewatering.}

\item{alevin_int}{A count of the number of intervals alevins can survive dewatering.}
}
\value{
A number of the proportion surviving.
}
\description{
Survival by Elevation and Intervals
}
