% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect-elevation.R
\name{edw_transect_elevation}
\alias{edw_transect_elevation}
\title{Transect Elevation}
\usage{
edw_transect_elevation(hydrograph, model, discharge = "Discharge",
  elevation = "Elevation")
}
\arguments{
\item{hydrograph}{A data frame of the discharge values.}

\item{model}{A data frame of discharge and elevation values.}

\item{discharge}{A string of the name of the column with the discharge values.}

\item{elevation}{A string of the name of the column with the elevation values.}
}
\value{
The hydrograph data with an elevation column.
}
\description{
Transect Elevation
}
\examples{
edw_transect_elevation(
  data.frame(Discharge = 12), 
  data.frame(Discharge = c(10,15), Elevation = c(40,50)))
}
