% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect-elevation2.R
\name{edw_transect_elevation2}
\alias{edw_transect_elevation2}
\title{Transect Elevation2}
\usage{
edw_transect_elevation2(hydrograph, model, discharge = c("Discharge",
  "Discharge2"), elevation = "Elevation")
}
\arguments{
\item{hydrograph}{A data frame of the discharge values.}

\item{model}{A data frame of discharge and elevation values.}

\item{discharge}{A string of the names of the column with the discharge values.}

\item{elevation}{A string of the name of the column with the elevation values.}
}
\value{
The hydrograph data with an elevation column.
}
\description{
Transect Elevation2
}
