% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect-weighting.R
\name{edw_transect_weighting}
\alias{edw_transect_weighting}
\title{Transect Weighting}
\usage{
edw_transect_weighting(rkm, data, riverkm = "RiverKm",
  spawning = "Spawning")
}
\arguments{
\item{rkm}{A double vector of the transect river kms}

\item{data}{A data frame with columns indicating the river km and proportion of spawning.}

\item{riverkm}{A string of the name of the column in data with the river km.}

\item{spawning}{A string of the name of the column in data with the spawning proportions.}
}
\value{
A double vector of the transect weightings.
}
\description{
Transect Weighting
}
\examples{
edw_transect_weighting(c(1,2), data = data.frame(
  RiverKm = c(0.5,1.9,3),
  Spawning = c(0.1,0.8,0.1)
))
}
