% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers_model}
\alias{rivers_model}
\title{A tibble of 1D model output for two rivers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 3 columns.}
\usage{
rivers_model
}
\description{
A tibble of 1D model output for two rivers
}
\keyword{datasets}
